Extract unmanaged operation handling
